Migrate to versioned module the config methods
