+++ chains/solana/contracts/programs/ccip-router/src/state.rs
@@ -1,7 +1,5 @@
 use anchor_lang::prelude::*;
 
-use crate::CcipRouterError;
-
 // zero_copy is used to prevent hitting stack/heap memory limits
 #[account(zero_copy)]
 #[derive(InitSpace, AnchorSerialize, AnchorDeserialize)]
@@ -230,7 +228,6 @@ impl TryFrom<u128> for MessageExecutionState {
 
 #[cfg(test)]
 mod tests {
-    use crate::CcipRouterError;
 
     use super::*;
     use std::convert::TryFrom;
@@ -256,242 +253,4 @@ mod tests {
         assert!(MessageExecutionState::try_from(4).is_err());
     }
 
-    #[test]
-    fn test_is_valid_on_ramp() {
-        struct SourceChainTestCase {
-            config: SourceChainConfig,
-            expected: bool,
-        }
-
-        let cases = vec![
-            SourceChainTestCase {
-                config: SourceChainConfig {
-                    on_ramp: vec![],
-                    is_enabled: true,
-                },
-                expected: true,
-            },
-            SourceChainTestCase {
-                config: SourceChainConfig {
-                    on_ramp: vec![0; 64],
-                    is_enabled: true,
-                },
-                expected: true,
-            },
-            SourceChainTestCase {
-                config: SourceChainConfig {
-                    on_ramp: vec![0; 128],
-                    is_enabled: true,
-                },
-                expected: true,
-            },
-            SourceChainTestCase {
-                config: SourceChainConfig {
-                    on_ramp: vec![0; 1],
-                    is_enabled: true,
-                },
-                expected: false,
-            },
-            SourceChainTestCase {
-                config: SourceChainConfig {
-                    on_ramp: vec![0; 63],
-                    is_enabled: true,
-                },
-                expected: false,
-            },
-            SourceChainTestCase {
-                config: SourceChainConfig {
-                    on_ramp: vec![0; 65],
-                    is_enabled: true,
-                },
-                expected: false,
-            },
-            SourceChainTestCase {
-                config: SourceChainConfig {
-                    on_ramp: vec![0; 127],
-                    is_enabled: true,
-                },
-                expected: false,
-            },
-            SourceChainTestCase {
-                config: SourceChainConfig {
-                    on_ramp: vec![0; 129],
-                    is_enabled: true,
-                },
-                expected: false,
-            },
-        ];
-
-        for case in cases {
-            let result = case.config.is_valid_on_ramp();
-            assert_eq!(
-                result,
-                case.expected,
-                "Failed for on_ramp length: {}",
-                case.config.on_ramp.len()
-            );
-        }
-    }
-
-    #[test]
-    fn test_get_on_ramps() {
-        struct TestCase {
-            on_ramp: Vec<u8>,
-            expected: Result<Vec<[u8; 64]>>,
-        }
-
-        let cases = vec![
-            TestCase {
-                on_ramp: vec![],
-                expected: Ok(Vec::new()),
-            },
-            TestCase {
-                on_ramp: vec![1; 64],
-                expected: Ok(vec![[1; 64]]),
-            },
-            TestCase {
-                on_ramp: [vec![2; 64], vec![3; 64]].concat(),
-                expected: Ok(vec![[2; 64], [3; 64]]),
-            },
-            TestCase {
-                on_ramp: vec![3; 1],
-                expected: Err(CcipRouterError::InvalidInputs.into()),
-            },
-            TestCase {
-                on_ramp: vec![4; 63],
-                expected: Err(CcipRouterError::InvalidInputs.into()),
-            },
-            TestCase {
-                on_ramp: vec![5; 65],
-                expected: Err(CcipRouterError::InvalidInputs.into()),
-            },
-            TestCase {
-                on_ramp: vec![6; 127],
-                expected: Err(CcipRouterError::InvalidInputs.into()),
-            },
-            TestCase {
-                on_ramp: vec![7; 129],
-                expected: Err(CcipRouterError::InvalidInputs.into()),
-            },
-        ];
-
-        for case in cases {
-            let config = SourceChainConfig {
-                is_enabled: true,
-                on_ramp: case.on_ramp,
-            };
-            let result = config.get_on_ramps();
-            assert_eq!(
-                result,
-                case.expected,
-                "Failed for on_ramp length: {}",
-                config.on_ramp.len()
-            );
-        }
-    }
-
-    #[test]
-    fn test_is_on_ramp_configured() {
-        struct ConfigTestCase {
-            config_on_ramps: Vec<u8>,
-            given_on_ramp: Vec<u8>,
-            expected: bool,
-        }
-
-        let empty_config = vec![];
-        let one_address_config = vec![1; 64];
-        let two_addresses_config = [vec![2; 64], vec![3; 64]].concat();
-        let two_smaller_addresses_config =
-            [vec![4; 32], vec![0; 32], vec![5; 32], vec![0; 32]].concat(); // TODO: Check if this is ok
-
-        let cases = vec![
-            // No address configured
-            ConfigTestCase {
-                config_on_ramps: empty_config.clone(),
-                given_on_ramp: vec![],
-                expected: false,
-            },
-            ConfigTestCase {
-                config_on_ramps: empty_config.clone(),
-                given_on_ramp: vec![1; 64],
-                expected: false,
-            },
-            ConfigTestCase {
-                config_on_ramps: empty_config.clone(),
-                given_on_ramp: vec![4; 64],
-                expected: false,
-            },
-            // One address configured
-            ConfigTestCase {
-                config_on_ramps: one_address_config.clone(),
-                given_on_ramp: vec![],
-                expected: false,
-            },
-            ConfigTestCase {
-                config_on_ramps: one_address_config.clone(),
-                given_on_ramp: vec![1; 64],
-                expected: true,
-            },
-            ConfigTestCase {
-                config_on_ramps: one_address_config.clone(),
-                given_on_ramp: vec![4; 64],
-                expected: false,
-            },
-            // Two addresses configured
-            ConfigTestCase {
-                config_on_ramps: two_addresses_config.clone(),
-                given_on_ramp: vec![],
-                expected: false,
-            },
-            ConfigTestCase {
-                config_on_ramps: two_addresses_config.clone(),
-                given_on_ramp: vec![2; 64],
-                expected: true,
-            },
-            ConfigTestCase {
-                config_on_ramps: two_addresses_config.clone(),
-                given_on_ramp: vec![3; 64],
-                expected: true,
-            },
-            ConfigTestCase {
-                config_on_ramps: two_addresses_config.clone(),
-                given_on_ramp: [vec![2; 32], vec![3; 32]].concat(),
-                expected: false,
-            },
-            // Two addresses configured with smaller addresses
-            ConfigTestCase {
-                config_on_ramps: two_smaller_addresses_config.clone(),
-                given_on_ramp: vec![],
-                expected: false,
-            },
-            ConfigTestCase {
-                config_on_ramps: two_smaller_addresses_config.clone(),
-                given_on_ramp: vec![0; 32],
-                expected: false,
-            },
-            ConfigTestCase {
-                config_on_ramps: two_smaller_addresses_config.clone(),
-                given_on_ramp: vec![4; 32],
-                expected: true,
-            },
-            ConfigTestCase {
-                config_on_ramps: two_smaller_addresses_config.clone(),
-                given_on_ramp: vec![5; 32],
-                expected: true,
-            },
-        ];
-
-        for case in cases {
-            let config = SourceChainConfig {
-                is_enabled: true,
-                on_ramp: case.config_on_ramps,
-            };
-            let result = config.is_on_ramp_configured(&case.given_on_ramp);
-            assert_eq!(
-                result, case.expected,
-                "Failed for on_ramp: {:?}",
-                case.given_on_ramp
-            );
-        }
-    }
 }
